requirement.txt
streamlit==1.49.1
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-chroma==0.2.6
langchain==0.3.27
huggingface-hub==0.35.0
google-genai==1.26.0
yarl==1.20.1
langgraph==0.6.8
langchain-google-genai